Move helper methods into the job files and register the jobs at controller startup.
